Move RoiMiniBatch to MTImageFeatureToBatch.scala
